# src/fk-bff-gerente/openapi.yaml
openapi: 3.1.0
info:
  title: fk-bff-gerente
  description: "Backend-for-frontend do GerenteRelacionamento. Fronteira web -- nao\
    \ e bounded context (CONTEXT-MAP.md). Conduz Authorization Code + PKCE + OIDC\
    \ contra fk-servidor-autorizacao, mantem a sessao em Redis via Spring Session\
    \ e expoe ao app-gerente somente o que o Angular precisa: identidade da sessao\
    \ corrente, a carteira, e a composicao da tela de atendimento a partir de fk-servico-carteira-clientes\
    \ e fk-servico-credito (AC30). Nenhum access token ou refresh token atravessa\
    \ esta fronteira para o browser (AC19) -- o BFF obtem por Token Exchange um token\
    \ para cada destino e nunca reencaminha o token de login (ADR-0015). Nao implementa\
    \ regra de credito, nao fala com fk-simulador-core-legado e nao substitui a autorizacao\
    \ de recurso feita pelos servicos. Produz o seu PROPRIO envelope publico de erro\
    \ -- nunca repassa o ProblemDetail upstream verbatim."
  version: "0003"
servers:
- url: /bff
  description: "Origem publica atras do nginx (docker compose), TLS local."
...
paths:
...
  /api/sessao:
    get:
      tags:
      - sessao
      summary: Identidade do GerenteRelacionamento autenticado na sessao corrente
      description: "O suficiente para o Angular saber quem esta logado sem inspecionar\
        \ o cookie de sessao, que e HttpOnly e portanto inacessivel a JavaScript."
      operationId: consultarSessaoAtual
      responses:
        "200":
          description: Sessao autenticada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: "Sem sessao autenticada (sem cookie, ou cookie invalido/expirado)."
      security:
      - cookieSessao: []
...
  /api/carteira/clientes:
    get:
      tags:
      - carteira
      summary: Pagina da CarteiraClientes do gerente autenticado
      description: "Proxy autenticado para fk-servico-carteira-clientes: o BFF obtem\
        \ por Token Exchange um token com audience restrita a servico-carteira-clientes\
        \ (ADR-0015) e repassa a chamada. O BFF nunca fala com fk-simulador-core-legado\
        \ (AC30) nem reimplementa a autorizacao de recurso, que pertence ao servico\
        \ dono (ADR-0007)."
      operationId: listarClientesDaCarteira
      parameters:
      - name: pagina
        in: query
        description: Pagina zero-based.
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: tamanho
        in: query
        description: Tamanho de pagina. Limite e o mesmo de fk-servico-carteira-clientes.
        required: false
        schema:
          type: integer
          default: 20
          maximum: 100
          minimum: 1
      responses:
        "200":
          description: "Pagina da carteira do gerente autenticado. Corpo JSON opaco\
            \ de fk-servico-carteira-clientes, repassado sem tipagem no BFF."
          content:
            application/json:
              schema:
                type: object
                description: Corpo cru repassado sem desserializacao no BFF.
        "401":
          description: Sem sessao autenticada.
        "403":
          description: "Sem direito de atendimento atual, segundo fk-servico-carteira-clientes.\
            \ Como esse servico ainda nao publica codigo, esta recusa hoje surge tipicamente\
            \ como 502 DEPENDENCIA_INDISPONIVEL."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
        "502":
          description: "Resposta de fk-servico-carteira-clientes nao reconhecida pela\
            \ allow-list deste BFF (inclui hoje qualquer 403/404, ate aquele servico\
            \ publicar codigo)."
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
        "503":
          description: fk-servico-carteira-clientes indisponivel ou inalcancavel no
            momento.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EnvelopeErroPublico"
      security:
      - cookieSessao: []
...
components:
  schemas:
    EnvelopeErroPublico:
      type: object
      description: "Envelope publico de erro do BFF -- so status+codigo, nunca o ProblemDetail\
        \ (RFC 7807) de um Resource Server verbatim."
      properties:
        status:
          type: integer
          format: int32
          description: "Status HTTP, espelhado do valor devolvido na resposta."
          example: 403
        codigo:
          type: string
          description: Codigo estavel do dominio. Omitido (nao null explicito) quando
            o upstream nao publica um -- caso de fk-servico-carteira-clientes em 403/404.
          enum:
          - IDEMPOTENCY_KEY_AUSENTE
          - IDEMPOTENCY_KEY_INVALIDA
          - COMANDO_ILEGIVEL
          - IDENTIFICADOR_INVALIDO
          - SEM_DIREITO_DE_ATENDIMENTO
          - CONTA_NAO_ENCONTRADA
          - LIMITE_VIGENTE_DESATUALIZADO
          - SOLICITACAO_NAO_TERMINAL_EXISTENTE
          - IDEMPOTENCIA_EM_PROCESSAMENTO
          - COMANDO_INVALIDO
          - LIMITE_SOLICITADO_NAO_AUMENTA
          - IDEMPOTENCIA_FINGERPRINT_DIVERGENTE
          - DEPENDENCIA_INDISPONIVEL
    SessionResponse:
      type: object
      properties:
        gerenteId:
          type: string
          description: Identificador do GerenteRelacionamento autenticado (claim sub
            do id_token).
          example: gerente.a
...
  securitySchemes:
    cookieSessao:
      type: apiKey
      description: "Cookie de sessao opaco (Spring Session/Redis), HttpOnly, Secure,\
        \ SameSite=Lax. Nunca carrega access token ou refresh token -- eles ficam\
        \ no OAuth2AuthorizedClientRepository, server-side."
      name: SESSION
      in: cookie